+++ mdmp/group_analysis/inds/__init__.py
@@ -6,31 +6,30 @@ Import from here or from :mod:`mdmp.group_analysis` / :mod:`mdmp` re-exports::
     from mdmp.group_analysis.inds import vote_individual_structures, ISAggregationResult
 """
 
-from .aggregation import (
+from .pipeline import (
+    aggregate_individual_structures,
+    aggregate_with_options,
+    as_inds_mdm_view,
+    compute_individual_structure_consensus,
+    pool_conditional_filtered_states,
+    run_inds_global_beta_mc,
+    vote_individual_structures,
+)
+from .pooled_filtering import build_plot_filt_from_subjects
+from .refit import IndsRefitResult, refit_on_consensus
+from .results import (
     ConditionalEdgePosteriorResult,
     GlobalBetaMCResult,
     ISAggregatedMDMView,
     ISAggregateOptions,
     ISAggregationResult,
+    ISMDMViewOptions,
+    ISMonteCarloOptions,
     ISPlotAdapter,
-    IndsRefitResult,
+    ISVoteOptions,
     MCContributorMode,
     MCPosteriorSource,
     PoolingMode,
-    aggregate_individual_structures,
-    aggregate_with_options,
-    as_inds_mdm_view,
-    build_plot_filt_from_subjects,
-    compute_individual_structure_consensus,
-    pool_conditional_filtered_states,
-    refit_on_consensus,
-    run_inds_global_beta_mc,
-    vote_individual_structures,
-)
-from .results import (
-    ISMDMViewOptions,
-    ISMonteCarloOptions,
-    ISVoteOptions,
     merge_aggregate_options,
 )
 from .voting import ThresholdMode, repair_dag_to_acyclic, vote_edge_frequencies