+++ mdmp/plotting/animation.py
@@ -48,7 +48,7 @@ def plot_idag(
     matplotlib.animation.FuncAnimation
         Animation object.
     """
-    require_data_for_plot(mdm_object, plot_kw="plot_data=...")
+    require_data_for_plot(mdm_object, plot_kw="time_series=...")
     if distribution == "filt":
         require_filt_for_plot(mdm_object, plot_kw="plot_filt=...")
         mt_list = mdm_object.Filt['mt']